% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustPeak.R
\name{adjustPeak}
\alias{adjustPeak}
\title{Adjust Peak Intensities Based on CNV Data}
\usage{
adjustPeak(pp, cc, overlapRegion, eta = 0.5)
}
\arguments{
\item{pp}{Peak matrix with cells in columns.}

\item{cc}{CNV matrix with cells in columns.}

\item{overlapRegion}{DataFrame of overlapping regions}

\item{eta}{Adjustment factor (default=0.5)}
}
\value{
Adjusted peak matrix
}
\description{
Adjust Peak Intensities Based on CNV Data
}
