% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringMetric.R
\name{clusteringMetric}
\alias{clusteringMetric}
\title{Calculate Clustering Quality Metric}
\usage{
clusteringMetric(D, clu, disMethod = c("cosine", "hamming", "euclidean"))
}
\arguments{
\item{D}{Data matrix}

\item{clu}{Cluster assignments}

\item{disMethod}{Distance method ("cosine", "hamming", or "euclidean")}
}
\value{
Combined clustering quality index
}
\description{
Calculate Clustering Quality Metric
}
