% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectRegions.R
\name{intersectRegions}
\alias{intersectRegions}
\title{Find Intersecting Genomic Regions}
\usage{
intersectRegions(peaks, cnvs)
}
\arguments{
\item{peaks}{Character vector of peak regions (format "chr:start-end")}

\item{cnvs}{Character vector of CNV regions (format "chr:start-end")}
}
\value{
DataFrame of overlapping regions
}
\description{
Find Intersecting Genomic Regions
}
