% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAAS.R
\name{MAAS}
\alias{MAAS}
\title{Integrating cell=cell similarities to infer consensus latent factors through self-expressive matrix factorization}
\usage{
MAAS(A1, A2, A3, dims)
}
\arguments{
\item{A1}{First data layer matrix (e.g. cell-cell similarities based on chromatin accessibility)}

\item{A2}{Second data layer matrix (e.g. cell-cell similarities based on copy numbers)}

\item{A3}{Third omics data matrix (e.g. cell-cell similarities based on SNVs)}

\item{dims}{Vector of latent dimensions to test}
}
\value{
List containing factorization results and parameters
}
\description{
Integrating cell=cell similarities to infer consensus latent factors through self-expressive matrix factorization
}
