% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maas_example}
\alias{maas_example}
\title{MAAS Example Dataset}
\format{
A list with 3 components:
\describe{
  \item{Peak}{Peak intensity matrix}
  \item{CNV}{CNV matrix}
  \item{snv}{SNV regions}
}
}
\source{
Simulated data
}
\usage{
maas_example
}
\description{
Wxample dataset containing single-cell peaks, CNV and SNV data for integration.
}
\examples{
data(maas_example)
str(maas_example)
}
\keyword{datasets}
